+++ core/embed/projects/bootloader_ci/messages.c
@@ -622,7 +622,7 @@ int process_msg_FirmwareUpload(uint8_t iface_num, uint32_t msg_size,
   const uint32_t *const src = (const uint32_t *const)chunk_buffer;
 
   ensure((chunk_size % FLASH_BLOCK_SIZE == 0) * sectrue, NULL);
-  for (int i = 0; i < chunk_size / FLASH_BLOCK_SIZE; i++) {
+  for (size_t i = 0; i < chunk_size / FLASH_BLOCK_SIZE; i++) {
     ensure(flash_area_write_block(
                &FIRMWARE_AREA,
                firmware_block * IMAGE_CHUNK_SIZE + i * FLASH_BLOCK_SIZE,
